Started refactoring code to unify all frameworks. Doesn't work yet
