+++ BatchProvider.lua
@@ -1,32 +1,4 @@
-local BatchProvider = torch.class('nnf.BatchProvider')
-
-local function createWindowBase(rec,i,j,is_bg)
-  local label = is_bg == true and 0+1 or rec.label[j]+1
-  local window = {i,rec.boxes[j][1],rec.boxes[j][2],
-                    rec.boxes[j][3],rec.boxes[j][4],
-                    label}
-  return window
-end
-
-local function createWindowAngle(rec,i,j,is_bg)
-  local label = is_bg == true and 0+1 or rec.label[j]+1
-  --local ang = ( is_bg == false and rec.objects[rec.correspondance[j] ] ) and 
-  --                  rec.objects[rec.correspondance[j] ].viewpoint.azimuth or 0
-  local ang
-  if is_bg == false and rec.objects[rec.correspondance[j] ] then
-    if rec.objects[rec.correspondance[j] ].viewpoint.distance == '0' then
-      ang = rec.objects[rec.correspondance[j] ].viewpoint.azimuth_coarse
-    else
-      ang = rec.objects[rec.correspondance[j] ].viewpoint.azimuth
-    end
-  else
-    ang = 0
-  end
-  local window = {i,rec.boxes[j][1],rec.boxes[j][2],
-                    rec.boxes[j][3],rec.boxes[j][4],
-                    label,ang}
-  return window
-end
+local BatchProvider = torch.class('nnf.BatchProvider','nnf.BatchProviderBase')
 
 function BatchProvider:__init(feat_provider)
   self.dataset = feat_provider.dataset
@@ -35,74 +7,10 @@ function BatchProvider:__init(feat_provider)
   self.nTimesMoreData = 10
   self.iter_per_batch = 500
   
-  self.batch_size = 128
-  self.fg_fraction = 0.25
-  
-  self.fg_threshold = 0.5
-  self.bg_threshold = {0.0,0.5}
-  
-  self.createWindow = createWindowBase--createWindowAngle
-  
   self.batch_dim = {256*50}
   self.target_dim = 1
   
-  self.do_flip = true
-  
-  --self:setupData()
 end
-
-
-function BatchProvider:setupData()
-  local dataset = self.dataset
-  local bb = {}
-  local bbT = {}
-
-  for i=0,dataset.num_classes do -- 0 because of background
-    bb[i] = {}
-  end
-
-  for i=1,dataset.num_imgs do
-    bbT[i] = {}
-  end
-
-  for i = 1,dataset.num_imgs do
-    if dataset.num_imgs > 10 then
-      xlua.progress(i,dataset.num_imgs)
-    end
-    
-    local rec = dataset:attachProposals(i)
-  
-    for j=1,rec:size() do    
-      local id = rec.label[j]
-      local is_fg = (rec.overlap[j] >= self.fg_threshold)
-      local is_bg = (not is_fg) and (rec.overlap[j] >= self.bg_threshold[1]  and
-                                     rec.overlap[j] <  self.bg_threshold[2])
-      if is_fg then
-        local window = self.createWindow(rec,i,j,is_bg)
-        table.insert(bb[1], window) -- could be id instead of 1
-      elseif is_bg then
-        local window = self.createWindow(rec,i,j,is_bg)
-        table.insert(bb[0], window)
-      end
-      
-    end
-    
-    for j=0,dataset.num_classes do -- 0 because of background
-      if #bb[j] > 0 then
-        bbT[i][j] = torch.FloatTensor(bb[j])
-      end
-    end
-        
-    bb = {}
-    for i=0,dataset.num_classes do -- 0 because of background
-      bb[i] = {}
-    end
-    collectgarbage()
-  end
-  self.bboxes = bbT
-  --return bbT
-end
-
 
 function BatchProvider:permuteIdx()
   local fg_num_each  = self.fg_num_each
@@ -234,7 +142,7 @@ function BatchProvider:prepareFeatures(im_idx,bboxes,fg_data,bg_data,fg_label,bg
 --  return fg_data,bg_data,fg_label,bg_label
 end
 
-function BatchProvider:getBatch(batches,targets)
+function BatchProvider:prepareBatch(batches,targets)
   local dataset = self.dataset
   
   self.fg_num_each = self.fg_fraction * self.batch_size
@@ -299,4 +207,22 @@ function BatchProvider:getBatch(batches,targets)
     end
   end
   return batches,targets
+end
+
+function BatchProvider:getBatch(batches,targets)
+  self._cur = self._cur or math.huge
+  -- we have reached the end of our batch pool, need to recompute
+  if self._cur > self.iter_per_batch then
+    self._batches,self._targets = self:prepareBatch(self._batches,self._targets)
+    self._cur = 1
+  end
+  --local batches = batches or torch.FloatTensor()
+  --local targets = targets or torch.FloatTensor() 
+
+  local batches = self._batches[self._cur]
+  local targets = self._targets[self._cur]
+  self._cur = self._cur + 1
+
+  return batches, targets
+
 end
